feedback_backend/pyproject.toml
[project]
name = "feedback-center"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==3.7.1",
    "bcrypt>=4.3.0",
    "cffi==1.17.1",
    "click==8.2.1",
    "cryptography==45.0.4",
    "ecdsa==0.19.1",
    "fastapi==0.104.1",
    "h11==0.16.0",
    "idna==3.10",
    "loguru==0.7.3",
    "passlib==1.7.4",
    "psycopg2-binary>=2.9.10",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic==2.4.2",
    "pydantic-core==2.10.1",
    "pydantic-settings==2.0.3",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.23",
    "starlette==0.27.0",
    "typing-extensions==4.14.0",
    "uvicorn==0.24.0",
]
